requirements.txt
requests
python-gnupg
beautifulsoup4
